added static_cast<int> to resolve unsigned_int errors

+++ RetroFE/Source/Collection/CollectionInfoBuilder.cpp
@@ -434,7 +434,7 @@ void CollectionInfoBuilder::addPlaylists(CollectionInfo *info)
         std::string basename = (std::string::npos == position)? file : file.substr(0, position);
 
         std::string comparator = ".txt";
-        int start = file.length() - comparator.length();
+        int start = static_cast<int>(file.length() - comparator.length());
 
         if(start >= 0)
         {
@@ -652,7 +652,7 @@ void CollectionInfoBuilder::ImportRomDirectory(std::string path, CollectionInfo 
                 for(extensionsIt = extensions.begin(); extensionsIt != extensions.end(); ++extensionsIt)
                 {
                     std::string comparator = "." + *extensionsIt;
-                    int start = file.length() - comparator.length() + 1;
+                    int start = static_cast<int>(file.length() - comparator.length()) + 1;
 
                     if (start >= 0)
                     {
